+++ crystalfontz/dbus/map/character.py
@@ -1,0 +1,11 @@
+from typing import ClassVar
+
+from crystalfontz.character import SpecialCharacter
+
+
+class SpecialCharacterM:
+    t: ClassVar[str] = "t"
+
+    @staticmethod
+    def unpack(character: int) -> SpecialCharacter:
+        raise NotImplementedError("load")
